requirements.txt
streamlit==1.35.0
torch==2.0.1
transformers==4.33.2
nltk==3.8.1
scikit-learn==1.4.1.post1
spacy==3.7.4
PyPDF2==3.0.1
fpdf2==2.7.8
langdetect==1.0.9
python-docx==1.1.0
numpy==1.26.0
pandas==2.1.1
